requirements.txt
# Core dependencies
streamlit>=1.28.0
google-generativeai>=0.3.0

# Utilities
python-dotenv>=1.0.0

# For future API integrations (commented for now)
requests>=2.31.0
aiohttp>=3.8.0